+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 2021-09-10
+### Added
+- Custom Widget to display top N slots and session attribute values
+### Changed
+- Bot tester now sets session attributes and sends audio utterances
+
 ## [0.2.0] - 2021-09-03
 ### Fixed
 - Changed metric namespace to include bot locale
@@ -20,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release
 
-[Unreleased]: https://github.com/aws-samples/aws-lex-v2-bot-analytics/compare/v0.2.0...develop
+[Unreleased]: https://github.com/aws-samples/aws-lex-v2-bot-analytics/compare/v0.3.0...develop
+[0.3.0]: https://github.com/aws-samples/aws-lex-v2-bot-analytics/releases/tag/v0.3.0
 [0.2.0]: https://github.com/aws-samples/aws-lex-v2-bot-analytics/releases/tag/v0.2.0
 [0.1.0]: https://github.com/aws-samples/aws-lex-v2-bot-analytics/releases/tag/v0.1.0